+++ src/main/java/com/github/hcsp/inheritance/Animal.java
@@ -1,0 +1,16 @@
+package com.github.hcsp.inheritance;
+
+public class Animal {
+    protected String name;
+
+    public Animal(String name) {
+        this.name = name;
+    }
+
+    public Animal() {
+    }
+
+    public void sayMyName(){
+        System.out.println("我的名字是" + name);
+    }
+}
